Add javadoc, move StandaloneActivity
