+++ hyperion-core/src/main/res/layout/activity_standalone.xml
